+++ app/src/main/kotlin/com/maltaisn/notes/model/DefaultPrefsManager.kt
@@ -191,6 +191,7 @@ class DefaultPrefsManager @Inject constructor(
         const val EXPORT_DATA = "export_data"
         const val AUTO_EXPORT = "auto_export"
         const val IMPORT_DATA = "import_data"
+        const val EXPORT_ARCHIVE = "export_archive"
         const val CLEAR_DATA = "clear_data"
         const val VERSION = "version"
         const val DELETED_TIMEOUT = "deleted_timeout"