+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added undo/redo support (#26).
 - Added Vietnamese translation (#146, @ngocanhtve) & updated other languages.
 - Allow adjusting the text size in the editor (#22).
+- Allow exporting notes as a ZIP archive (#59).
 - Added a setting for initially focusing on the content when a new note is created (#116).
 - Focus title if backspace is pressed at the start of content in a text note (#157).
 - Add new list item if enter is pressed in title and no items exist.